Updated build to gen messages file and refactored.
